Broadcom Android Release 8.1 Phase 17.4 (eng 1) code drop

+++ astra/linux/bcm_astra/tzioc_coredump.h
@@ -46,7 +46,7 @@
 int _tzioc_peer_coredump(
 	struct tzioc_client *pClient,
 	const char *pPeerName,
-	uint32_t ulPaddr,
+	uintptr_t ulPaddr,
     uint32_t ulBytes);
